# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nanocloud",
    "description": "A CLI tool for managing your Nanocloud",
    "license": {
      "name": ""
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/apis/nanocloud.io/v1/namespaces/{namespace}/bundles": {
      "get": {
        "tags": [
          "nanocloud"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Namespace to query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of entries to return",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "continue",
            "in": "query",
            "description": "Continue token for pagination",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bundle list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleList"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "NanocloudCertificate": []
          },
          {
            "NanocloudBearer": [
              "bundles.read"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "nanocloud"
        ],
        "operationId": "create",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Namespace for the new bundle",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bundle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Bundle created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bundle"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bundle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Bundle already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "NanocloudCertificate": []
          },
          {
            "NanocloudBearer": [
              "bundles.manage"
            ]
          }
        ]
      }
    },
    "/apis/nanocloud.io/v1/namespaces/{namespace}/bundles/{name}": {
      "get": {
        "tags": [
          "nanocloud"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Namespace of the bundle",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Bundle name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bundle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bundle"
                }
              }
            }
          },
          "404": {
            "description": "Bundle not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "NanocloudCertificate": []
          },
          {
            "NanocloudBearer": [
              "bundles.read"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "nanocloud"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "namespace",
            "in": "path",
            "description": "Namespace of the bundle",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Bundle name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Bundle deletion accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bundle"
                }
              }
            }
          },
          "404": {
            "description": "Bundle not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "NanocloudCertificate": []
          },
          {
            "NanocloudBearer": [
              "bundles.manage"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplyConflict": {
        "type": "object",
        "required": [
          "path",
          "existingManager"
        ],
        "properties": {
          "existingManager": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "BindingHistoryEntry": {
        "type": "object",
        "required": [
          "bindingId",
          "service",
          "status",
          "attempts"
        ],
        "properties": {
          "attempts": {
            "type": "integer",
            "format": "int32",
            "description": "Total attempts recorded for the binding.",
            "minimum": 0
          },
          "bindingId": {
            "type": "string"
          },
          "command": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "durationMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "eventId": {
            "type": [
              "string",
              "null"
            ]
          },
          "exit_code": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "lastFinishedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastStartedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "service": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/BindingHistoryStatus"
          },
          "stderr": {
            "type": [
              "string",
              "null"
            ]
          },
          "stdout": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "BindingHistoryStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "RUNNING",
          "SUCCEEDED",
          "FAILED",
          "TIMED_OUT"
        ]
      },
      "Bundle": {
        "type": "object",
        "required": [
          "apiVersion",
          "kind",
          "metadata",
          "spec"
        ],
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/ObjectMeta"
          },
          "spec": {
            "$ref": "#/components/schemas/BundleSpec"
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BundleStatus"
              }
            ]
          }
        }
      },
      "BundleCondition": {
        "type": "object",
        "required": [
          "type",
          "status"
        ],
        "properties": {
          "lastTransitionTime": {
            "type": [
              "string",
              "null"
            ],
            "description": "RFC3339 timestamp for the last transition."
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human readable context for the condition."
          },
          "reason": {
            "type": [
              "string",
              "null"
            ],
            "description": "Machine readable reason (e.g. `ProfileMissing`, `ApplySucceeded`)."
          },
          "status": {
            "$ref": "#/components/schemas/BundleConditionStatus",
            "description": "Overall status (True/False/Unknown)."
          },
          "type": {
            "$ref": "#/components/schemas/BundleConditionKind",
            "description": "Logical condition identifier (e.g. `Ready`, `Bound`)."
          }
        }
      },
...
      "BundleList": {
        "type": "object",
        "required": [
          "apiVersion",
          "kind",
          "metadata",
          "items"
        ],
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bundle"
            }
          },
          "kind": {
            "type": "string"
          },
          "metadata": {
            "$ref": "#/components/schemas/ListMeta"
          }
        }
      },
      "BundlePhase": {
        "type": "string",
        "enum": [
          "PENDING",
          "VALIDATING",
          "RENDERING",
          "APPLYING",
          "READY",
          "FAILED",
          "DELETING"
        ]
      },
...
      "BundleSecurityProfile": {
        "type": "object",
        "properties": {
          "allowPrivileged": {
            "type": "boolean",
            "description": "Gate for privileged capability requests (CAP_SYS_ADMIN, CAP_SYS_PTRACE, etc)."
          },
          "extraCapabilities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Explicit capabilities to inject into the container's allowed set."
          },
          "seccompProfile": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BundleSeccompProfile",
                "description": "Optional seccomp profile override."
              }
            ]
          }
        }
      },
      "BundleSnapshotSource": {
        "type": "object",
        "required": [
          "source"
        ],
        "properties": {
          "mediaType": {
            "type": [
              "string",
              "null"
            ],
            "description": "Media type for the snapshot artifact (default `application/x-tar`)."
          },
          "source": {
            "type": "string",
            "description": "URI referencing the snapshot payload (e.g. `keyspace://profiles/...` or `file:///path`)."
          }
        }
      },
      "BundleSpec": {
        "type": "object",
        "required": [
          "service"
        ],
        "properties": {
          "key": {
            "type": [
              "string",
              "null"
            ],
            "description": "Wrapped profile encryption key (base64-encoded)."
          },
          "namespace": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional namespace scope; defaults to `default` when omitted."
          },
          "options": {
            "type": "object",
            "description": "User-supplied installation options.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "security": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BundleSecurityProfile",
                "description": "Optional runtime security profile."
              }
            ]
          },
          "service": {
            "type": "string",
            "description": "Target service identifier (matching the OCI manifest name)."
          },
          "snapshot": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BundleSnapshotSource",
                "description": "Optional snapshot source used to seed persistent volumes at install time."
              }
            ]
          },
          "start": {
            "type": "boolean",
            "description": "Indicates whether the workload should be started after reconciliation."
          },
          "update": {
            "type": "boolean",
            "description": "Forces a fresh pull of the service image even if cached layers already exist."
          }
        }
      },
      "BundleStatus": {
        "type": "object",
        "properties": {
          "bindingHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BindingHistoryEntry"
            },
            "description": "Latest binding execution outcomes."
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BundleCondition"
            },
            "description": "Detailed condition set mirroring Kubernetes conventions."
          },
          "lastReconciledTime": {
            "type": [
              "string",
              "null"
            ],
            "description": "Timestamp of the last successful reconciliation."
          },
          "observed_generation": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Observed bundle generation."
          },
          "phase": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BundlePhase",
                "description": "High-level reconciliation phase."
              }
            ]
          },
          "workload": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BundleWorkloadRef",
                "description": "Workload reference emitted by the controller (StatefulSet/Deployment/etc)."
              }
            ]
          }
        }
      },
      "BundleWorkloadRef": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Workload name materialised from the bundle."
          },
          "namespace": {
            "type": [
              "string",
              "null"
            ],
            "description": "Namespace hosting the workload; defaults to the bundle namespace when omitted."
          },
          "uid": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional UID of the managed resource."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "conflicts": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/ApplyConflict"
            }
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "ListMeta": {
        "type": "object",
        "description": "Metadata included with Kubernetes list responses.",
        "properties": {
          "continue": {
            "type": [
              "string",
              "null"
            ]
          },
          "remainingItemCount": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "resourceVersion": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ObjectMeta": {
        "type": "object",
        "description": "Minimal representation of Kubernetes object metadata.",
        "properties": {
          "annotations": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "labels": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "namespace": {
            "type": [
              "string",
              "null"
            ]
          },
          "resourceVersion": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}